+++ apps/admin/components/admin/Actions.tsx
@@ -1,0 +1,48 @@
+const positions = [
+  {
+    id: 1,
+    title: "Approved",
+    type: "✓",
+    location: "Remote",
+    department: "Engineering",
+    closeDate: "2020-01-07",
+    closeDateFull: "January 7, 2020",
+  },
+  {
+    id: 2,
+    title: "RNG",
+    type: "✓",
+    location: "Remote",
+    department: "Engineering",
+    closeDate: "2020-01-07",
+    closeDateFull: "January 7, 2020",
+  },
+];
+
+// mostly dummy data for now
+export const Actions = () => {
+  return (
+    <div className="bg-white shadow overflow-hidden sm:rounded-md">
+      <ul role="list" className="divide-y divide-gray-200">
+        {positions.map((position) => (
+          <li key={position.id}>
+            <a href="#" className="block hover:bg-gray-50">
+              <div className="px-4 py-4 sm:px-6">
+                <div className="flex items-center justify-between">
+                  <p className="text-sm font-medium text-indigo-600 truncate">
+                    {position.title}
+                  </p>
+                  <div className="ml-2 flex-shrink-0 flex">
+                    <p className="px-2 inline-flex text-xs leading-5 font-semibold rounded-full bg-green-100 text-green-800">
+                      {position.type}
+                    </p>
+                  </div>
+                </div>
+              </div>
+            </a>
+          </li>
+        ))}
+      </ul>
+    </div>
+  );
+};
