+++ apps/admin/components/app/Hero.module.css
@@ -1,0 +1,17 @@
+.gradient {
+  background-size: 400%;
+
+  -webkit-animation: Gradient 5s ease infinite;
+  -moz-animation: Gradient 5s ease infinite;
+  animation: Gradient 5s ease infinite;
+}
+
+@keyframes Gradient {
+  0%,
+  100% {
+    background-position: 0% 50%;
+  }
+  50% {
+    background-position: 100% 50%;
+  }
+}
