+++ apps/admin/tailwind.config.js
@@ -3,9 +3,18 @@ module.exports = {
     "./pages/**/*.{js,ts,jsx,tsx}",
     "./components/**/*.{js,ts,jsx,tsx}",
   ],
-  theme: { extend: {} },
-  plugins: [
-    require("@tailwindcss/forms"),
-    // ...
-  ],
+  theme: {
+    extend: {
+      keyframes: {
+        gradient: {
+          "0%, 100%": { "background-position": "0% 50%" },
+          "50%": { "background-position": "100% 50%;" },
+        },
+      },
+      animation: {
+        gradient: "gradient 1s infinite",
+      },
+    },
+  },
+  plugins: [require("@tailwindcss/forms")],
 };
